pyproject.toml
[project]
name = "my-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "kafka-python>=2.3.0",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
]
